+++ src/Resources/contao/config/config.php
@@ -141,7 +141,7 @@ if (TL_MODE == 'BE' && strpos(Environment::get('request'), 'contao/install') ===
     $GLOBALS['TL_HOOKS']['outputBackendTemplate'][] = ['OMOSde\ContaoOmBackendBundle\Hooks', 'addBodyClasses'];
     $GLOBALS['TL_HOOKS']['outputBackendTemplate'][] = ['OMOSde\ContaoOmBackendBundle\BackendLinks', 'addBackendLinks'];
     $GLOBALS['TL_HOOKS']['outputBackendTemplate'][] = ['OMOSde\ContaoOmBackendBundle\Versions', 'checkContaoVersions'];
-    //$GLOBALS['TL_HOOKS']['postLogin'][] = ['OMOSde\ContaoOmBackendBundle\Hooks', 'redirectUser'];
+    $GLOBALS['TL_HOOKS']['postLogin'][] = ['OMOSde\ContaoOmBackendBundle\Hooks', 'redirectUser'];
     $GLOBALS['TL_HOOKS']['getUserNavigation'][] = ['OMOSde\ContaoOmBackendBundle\Hooks', 'handleModuleOrder'];
     //$GLOBALS['TL_HOOKS']['outputBackendTemplate'][] = array('OMOSde\ContaoOmBackendBundle\BackendLinks', 'addBackendLinksMain');
     $GLOBALS['TL_HOOKS']['getUserNavigation'][] = ['OMOSde\ContaoOmBackendBundle\ModuleBackendTabs', 'changeNavigation'];